requirements.txt
Flask==2.3.2
PyPDF2==2.10.3
pdf2image==1.16.0
Pillow==9.3.0
pytesseract==0.3.10